Separated MainNative and MainTestNative
