+++ platformio.ini
@@ -54,6 +54,7 @@ lib_ignore =
     ArduinoNative
     HALSim
     HALTest
+    MainNative
 extra_scripts =
 monitor_speed = 115200
 monitor_filters = esp32_exception_decoder
@@ -81,6 +82,7 @@ lib_deps =
     ArduinoNative
     HALSim
     Webots
+    MainNative
 lib_ignore =
     HALTarget
     HALTest
@@ -110,6 +112,7 @@ lib_deps =
 lib_ignore =
     HALSim
     HALTarget
+    MainNative
 extra_scripts =
 
 ; *****************************************************************************
